Add test for processing before describe
Also refactor existing processing tests to carry out actual processing instead of only using mocks
