+++ utils/tests/tests.py
@@ -464,279 +464,3 @@ class BulkDescribeUtils(TestCase):
         new_sound2 = Sound.objects.get(user=user, original_filename='file5.wav')  # New version of last correct sound
         self.assertNotEqual(new_sound1.id, sound1_id)  # Check that IDs are not the same
         self.assertNotEqual(new_sound2.id, sound2_id)  # Check that IDs are not the same
-
-
-def convert_to_pcm_mock(input_filename, output_filename):
-    return True
-
-
-def convert_to_pcm_mock_create_file(input_filename, output_filename):
-    create_test_files(paths=[output_filename], n_bytes=2048)
-    return True
-
-
-def convert_to_pcm_mock_fail(input_filename, output_filename):
-    raise AudioProcessingException("failed converting to pcm")
-
-
-def stereofy_mock(stereofy_executble_path, input_filename, output_filename):
-    return dict(
-        duration=123.5,
-        channels=2,
-        samplerate=44100,
-        bitdepth=16)
-
-
-def stereofy_mock_fail(stereofy_executble_path, input_filename, output_filename):
-    raise AudioProcessingException("stereofy has failed")
-
-
-def convert_to_mp3_mock(input_filename, output_filename, quality):
-    create_test_files(paths=[output_filename])
-
-
-def convert_to_mp3_mock_fail(input_filename, output_filename, quality):
-    raise AudioProcessingException("conversion to mp3 (preview) has failed")
-
-
-def convert_to_ogg_mock(input_filename, output_filename, quality):
-    create_test_files(paths=[output_filename])
-
-
-def convert_to_ogg_mock_fail(input_filename, output_filename, quality):
-    raise AudioProcessingException("conversion to ogg (preview) has failed")
-
-
-def create_wave_images_mock(
-        input_filename, output_filename_w, output_filename_s, image_width, image_height, fft_size, **kwargs):
-    create_test_files(paths=[output_filename_w, output_filename_s])
-
-
-def create_wave_images_mock_fail(
-        input_filename, output_filename_w, output_filename_s, image_width, image_height, fft_size, **kwargs):
-    raise AudioProcessingException("creation of display images has failed")
-
-
-class AudioProcessingTestCase(TestCase):
-
-    fixtures = ['licenses']
-
-    def pre_test(self, create_sound_file=True):
-        # Do some stuff which needs to be carried out right before each test
-        self.assertEqual(self.sound.processing_state, "PE")
-        if create_sound_file:
-            create_test_files(paths=[self.sound.locations('path')])  # Create fake file for original sound path
-
-    def setUp(self):
-        user, _, sounds = create_user_and_sounds(num_sounds=1, type="mp3")  # Use mp3 so it needs converstion to PCM
-        self.sound = sounds[0]
-        self.user = user
-
-    def test_sound_object_does_not_exist(self):
-        with self.assertRaises(AudioProcessingException) as cm:
-            FreesoundAudioProcessor(sound_id=999)
-        exc = cm.exception
-        self.assertIn('did not find Sound object', exc.message)
-
-    @override_settings(USE_PREVIEWS_WHEN_ORIGINAL_FILES_MISSING=False)
-    @override_processing_tmp_path_with_temp_directory
-    def test_sound_path_does_not_exist(self):
-        self.pre_test(create_sound_file=False)
-        result = FreesoundAudioProcessor(sound_id=Sound.objects.first().id).process()
-        self.assertFalse(result)  # Processing failed, retutned False
-        self.sound.refresh_from_db()
-        self.assertEqual(self.sound.processing_state, "FA")
-        self.assertEqual(self.sound.processing_ongoing_state, "FI")
-        self.assertIn('could not find file with path', self.sound.processing_log)
-        self.assertFalse(len(os.listdir(settings.PROCESSING_TEMP_DIR)), 0)
-
-    @mock.patch('utils.audioprocessing.processing.convert_to_pcm', side_effect=convert_to_pcm_mock_fail)
-    @override_settings(USE_PREVIEWS_WHEN_ORIGINAL_FILES_MISSING=False)
-    @override_processing_tmp_path_with_temp_directory
-    @override_sounds_path_with_temp_directory
-    def test_conversion_to_pcm_failed(self, *args):
-        self.pre_test()
-        result = FreesoundAudioProcessor(sound_id=Sound.objects.first().id).process()
-        # will fail because mocked version of convert_to_pcm fails
-        self.assertFalse(result)  # Processing failed, retutned False
-        self.sound.refresh_from_db()
-        self.assertEqual(self.sound.processing_state, "FA")
-        self.assertEqual(self.sound.processing_ongoing_state, "FI")
-        self.assertIn('conversion to PCM failed', self.sound.processing_log)
-        self.assertFalse(len(os.listdir(settings.PROCESSING_TEMP_DIR)), 0)
-
-    @override_settings(USE_PREVIEWS_WHEN_ORIGINAL_FILES_MISSING=False)
-    @override_processing_tmp_path_with_temp_directory
-    @override_sounds_path_with_temp_directory
-    def test_no_need_to_convert_to_pcm(self):
-        self.sound.type = 'wav'
-        self.sound.save()
-        self.pre_test()
-        result = FreesoundAudioProcessor(sound_id=Sound.objects.first().id).process()
-        self.assertFalse(result)  # Processing failed, retutned False
-        self.sound.refresh_from_db()
-        self.assertEqual(self.sound.processing_state, "FA")
-        self.assertEqual(self.sound.processing_ongoing_state, "FI")
-        self.assertIn('no need to convert, this file is already PCM data', self.sound.processing_log)
-        self.assertFalse(len(os.listdir(settings.PROCESSING_TEMP_DIR)), 0)
-        # NOTE: this test will generate stereofy errors as well but here we only check that a spcific message about
-        # PCM conversion was added to the log. stereofy is tested below.
-
-    @mock.patch('utils.audioprocessing.processing.stereofy_and_find_info', side_effect=stereofy_mock_fail)
-    @mock.patch('utils.audioprocessing.processing.convert_to_pcm', side_effect=convert_to_pcm_mock)
-    @override_settings(USE_PREVIEWS_WHEN_ORIGINAL_FILES_MISSING=False)
-    @override_processing_tmp_path_with_temp_directory
-    @override_sounds_path_with_temp_directory
-    def test_stereofy_failed(self, *args):
-        self.pre_test()
-        result = FreesoundAudioProcessor(sound_id=Sound.objects.first().id).process()
-        # processing will fail because stereofy mock raises an exception
-        self.assertFalse(result)  # Processing failed, retutned False
-        self.sound.refresh_from_db()
-        self.assertEqual(self.sound.processing_state, "FA")
-        self.assertEqual(self.sound.processing_ongoing_state, "FI")
-        self.assertIn('stereofy has failed', self.sound.processing_log)
-        self.assertFalse(len(os.listdir(settings.PROCESSING_TEMP_DIR)), 0)
-
-    @mock.patch('utils.audioprocessing.processing.stereofy_and_find_info', side_effect=stereofy_mock)
-    @mock.patch('utils.audioprocessing.processing.convert_to_pcm', side_effect=convert_to_pcm_mock)
-    @override_settings(USE_PREVIEWS_WHEN_ORIGINAL_FILES_MISSING=False)
-    @override_processing_tmp_path_with_temp_directory
-    @override_sounds_path_with_temp_directory
-    def test_set_audio_info_fields(self, *args):
-        self.pre_test()
-        FreesoundAudioProcessor(sound_id=Sound.objects.first().id).process()
-        self.sound.refresh_from_db()
-        self.assertEqual(self.sound.duration, 123.5)  # Assert that info properties were set
-        self.assertEqual(self.sound.channels, 2)
-        self.assertEqual(self.sound.samplerate, 44100)
-        self.assertEqual(self.sound.bitrate, 0)
-        self.assertEqual(self.sound.bitdepth, 0)  # This will be 0 because sound is mp3 and bitdepth is overwritten to 0
-        # NOTE: after calling set_audio_info_fields processing will fail, but we're only interested in testing up to
-        # this point for the present unit test
-
-    @mock.patch('utils.audioprocessing.processing.convert_to_mp3', side_effect=convert_to_mp3_mock_fail)
-    @mock.patch('utils.audioprocessing.processing.stereofy_and_find_info', side_effect=stereofy_mock)
-    @mock.patch('utils.audioprocessing.processing.convert_to_pcm', side_effect=convert_to_pcm_mock)
-    @override_settings(USE_PREVIEWS_WHEN_ORIGINAL_FILES_MISSING=False)
-    @override_processing_tmp_path_with_temp_directory
-    @override_sounds_path_with_temp_directory
-    @override_previews_path_with_temp_directory
-    def test_make_mp3_previews_fails(self, *args):
-        self.pre_test()
-        result = FreesoundAudioProcessor(sound_id=Sound.objects.first().id).process()
-        # processing will fail because convert_to_mp3 mock raises an exception
-        self.assertFalse(result)  # Processing failed, retutned False
-        self.sound.refresh_from_db()
-        self.assertEqual(self.sound.processing_state, "FA")
-        self.assertEqual(self.sound.processing_ongoing_state, "FI")
-        self.assertIn('conversion to mp3 (preview) has failed', self.sound.processing_log)
-        self.assertFalse(len(os.listdir(settings.PROCESSING_TEMP_DIR)), 0)
-
-    @mock.patch('utils.audioprocessing.processing.convert_to_ogg', side_effect=convert_to_ogg_mock_fail)
-    @mock.patch('utils.audioprocessing.processing.convert_to_mp3', side_effect=convert_to_mp3_mock)
-    @mock.patch('utils.audioprocessing.processing.stereofy_and_find_info', side_effect=stereofy_mock)
-    @mock.patch('utils.audioprocessing.processing.convert_to_pcm', side_effect=convert_to_pcm_mock)
-    @override_settings(USE_PREVIEWS_WHEN_ORIGINAL_FILES_MISSING=False)
-    @override_processing_tmp_path_with_temp_directory
-    @override_sounds_path_with_temp_directory
-    @override_previews_path_with_temp_directory
-    def test_make_ogg_previews_fails(self, *args):
-        self.pre_test()
-        result = FreesoundAudioProcessor(sound_id=Sound.objects.first().id).process()
-        # processing will fail because convert_to_ogg mock raises an exception
-        self.assertFalse(result)  # Processing failed, retutned False
-        self.sound.refresh_from_db()
-        self.assertEqual(self.sound.processing_state, "FA")
-        self.assertEqual(self.sound.processing_ongoing_state, "FI")
-        self.assertIn('conversion to ogg (preview) has failed', self.sound.processing_log)
-        self.assertFalse(len(os.listdir(settings.PROCESSING_TEMP_DIR)), 0)
-
-    @mock.patch('utils.audioprocessing.processing.create_wave_images', side_effect=create_wave_images_mock_fail)
-    @mock.patch('utils.audioprocessing.processing.convert_to_ogg', side_effect=convert_to_ogg_mock)
-    @mock.patch('utils.audioprocessing.processing.convert_to_mp3', side_effect=convert_to_mp3_mock)
-    @mock.patch('utils.audioprocessing.processing.stereofy_and_find_info', side_effect=stereofy_mock)
-    @mock.patch('utils.audioprocessing.processing.convert_to_pcm', side_effect=convert_to_pcm_mock)
-    @override_settings(USE_PREVIEWS_WHEN_ORIGINAL_FILES_MISSING=False)
-    @override_processing_tmp_path_with_temp_directory
-    @override_sounds_path_with_temp_directory
-    @override_previews_path_with_temp_directory
-    @override_displays_path_with_temp_directory
-    def test_create_images_fails(self, *args):
-        self.pre_test()
-        result = FreesoundAudioProcessor(sound_id=Sound.objects.first().id).process()
-        # processing will fail because create_wave_images mock raises an exception
-        self.assertFalse(result)  # Processing failed, retutned False
-        self.sound.refresh_from_db()
-        self.assertEqual(self.sound.processing_state, "FA")
-        self.assertEqual(self.sound.processing_ongoing_state, "FI")
-        self.assertIn('creation of display images has failed', self.sound.processing_log)
-        self.assertFalse(len(os.listdir(settings.PROCESSING_TEMP_DIR)), 0)
-
-    @mock.patch('utils.audioprocessing.processing.create_wave_images', side_effect=create_wave_images_mock)
-    @mock.patch('utils.audioprocessing.processing.convert_to_ogg', side_effect=convert_to_ogg_mock)
-    @mock.patch('utils.audioprocessing.processing.convert_to_mp3', side_effect=convert_to_mp3_mock)
-    @mock.patch('utils.audioprocessing.processing.stereofy_and_find_info', side_effect=stereofy_mock)
-    @mock.patch('utils.audioprocessing.processing.convert_to_pcm', side_effect=convert_to_pcm_mock)
-    @override_settings(USE_PREVIEWS_WHEN_ORIGINAL_FILES_MISSING=False)
-    @override_processing_tmp_path_with_temp_directory
-    @override_sounds_path_with_temp_directory
-    @override_previews_path_with_temp_directory
-    @override_displays_path_with_temp_directory
-    def test_skip_previews(self, *args):
-        self.pre_test()
-        result = FreesoundAudioProcessor(sound_id=Sound.objects.first().id)\
-            .process(skip_previews=True)
-
-        self.assertFalse(os.path.exists(self.sound.locations('preview.LQ.ogg.path')))
-        self.assertFalse(os.path.exists(self.sound.locations('preview.HQ.ogg.path')))
-        self.assertFalse(os.path.exists(self.sound.locations('preview.LQ.mp3.path')))
-        self.assertFalse(os.path.exists(self.sound.locations('preview.HQ.mp3.path')))
-        self.assertTrue(os.path.exists(self.sound.locations('display.spectral.M.path')))
-        self.assertTrue(os.path.exists(self.sound.locations('display.spectral_bw.M.path')))
-        self.assertTrue(os.path.exists(self.sound.locations('display.spectral.L.path')))
-        self.assertTrue(os.path.exists(self.sound.locations('display.spectral_bw.L.path')))
-        self.assertTrue(os.path.exists(self.sound.locations('display.wave.M.path')))
-        self.assertTrue(os.path.exists(self.sound.locations('display.wave_bw.M.path')))
-        self.assertTrue(os.path.exists(self.sound.locations('display.wave.L.path')))
-        self.assertTrue(os.path.exists(self.sound.locations('display.wave_bw.L.path')))
-
-        self.assertTrue(result)  # Processing succeeded
-        self.sound.refresh_from_db()
-        self.assertEqual(self.sound.processing_state, "OK")
-        self.assertEqual(self.sound.processing_ongoing_state, "FI")
-        self.assertFalse(len(os.listdir(settings.PROCESSING_TEMP_DIR)), 0)
-
-    @mock.patch('utils.audioprocessing.processing.create_wave_images', side_effect=create_wave_images_mock)
-    @mock.patch('utils.audioprocessing.processing.convert_to_ogg', side_effect=convert_to_ogg_mock)
-    @mock.patch('utils.audioprocessing.processing.convert_to_mp3', side_effect=convert_to_mp3_mock)
-    @mock.patch('utils.audioprocessing.processing.stereofy_and_find_info', side_effect=stereofy_mock)
-    @mock.patch('utils.audioprocessing.processing.convert_to_pcm', side_effect=convert_to_pcm_mock)
-    @override_settings(USE_PREVIEWS_WHEN_ORIGINAL_FILES_MISSING=False)
-    @override_processing_tmp_path_with_temp_directory
-    @override_sounds_path_with_temp_directory
-    @override_previews_path_with_temp_directory
-    @override_displays_path_with_temp_directory
-    def test_skip_displays(self, *args):
-        self.pre_test()
-        result = FreesoundAudioProcessor(sound_id=Sound.objects.first().id) \
-            .process(skip_displays=True)
-
-        self.assertTrue(os.path.exists(self.sound.locations('preview.LQ.ogg.path')))
-        self.assertTrue(os.path.exists(self.sound.locations('preview.HQ.ogg.path')))
-        self.assertTrue(os.path.exists(self.sound.locations('preview.LQ.mp3.path')))
-        self.assertTrue(os.path.exists(self.sound.locations('preview.HQ.mp3.path')))
-        self.assertFalse(os.path.exists(self.sound.locations('display.spectral.M.path')))
-        self.assertFalse(os.path.exists(self.sound.locations('display.spectral_bw.M.path')))
-        self.assertFalse(os.path.exists(self.sound.locations('display.spectral.L.path')))
-        self.assertFalse(os.path.exists(self.sound.locations('display.spectral_bw.L.path')))
-        self.assertFalse(os.path.exists(self.sound.locations('display.wave.M.path')))
-        self.assertFalse(os.path.exists(self.sound.locations('display.wave_bw.M.path')))
-        self.assertFalse(os.path.exists(self.sound.locations('display.wave.L.path')))
-        self.assertFalse(os.path.exists(self.sound.locations('display.wave_bw.L.path')))
-
-        self.assertTrue(result)  # Processing succeeded
-        self.sound.refresh_from_db()
-        self.assertEqual(self.sound.processing_state, "OK")
-        self.assertEqual(self.sound.processing_ongoing_state, "FI")
-        self.assertFalse(len(os.listdir(settings.PROCESSING_TEMP_DIR)), 0)
